# lib/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bizny API - Productivity and venture coordination platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List venture templates
      parameters:
        - name: industry
          in: query
          schema:
            type: string
        - name: creatorId
          in: query
          schema:
            type: integer
        - name: visibility
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VentureTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new venture template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VentureTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentureTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentureTemplate"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a venture template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VentureTemplateInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentureTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a venture template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /templates/{id}/clone:
    post:
      operationId: cloneTemplate
      tags: [templates]
      summary: Clone a template to your library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Cloned template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentureTemplate"
  /templates/{id}/follow:
    post:
      operationId: followTemplate
      tags: [templates]
      summary: Follow a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
    delete:
      operationId: unfollowTemplate
      tags: [templates]
      summary: Unfollow a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /templates/{id}/save:
    post:
      operationId: saveTemplate
      tags: [templates]
      summary: Save a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
    delete:
      operationId: unsaveTemplate
      tags: [templates]
      summary: Unsave a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsaved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /templates/{id}/adopt:
    post:
      operationId: adoptTemplate
      tags: [templates]
      summary: Adopt a template — creates an execution instance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                instanceType:
                  type: string
                country:
                  type: string
                stateCity:
                  type: string
                startDate:
                  type: string
      responses:
        "201":
          description: Execution instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionInstance"
  /templates/{id}/fork:
    post:
      operationId: forkTemplate
      tags: [templates]
      summary: Fork a template to create a customised copy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        "201":
          description: Forked template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentureTemplate"
  /templates/{id}/status:
    get:
      operationId: getTemplateStatus
      tags: [templates]
      summary: Get follow/save status of a template for the current user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  following:
                    type: boolean
                  saved:
                    type: boolean
...
components:
  schemas:
...
    VentureTemplate:
      type: object
      required: [id, title, industry, description, durationDays, requiredSkills, requiredResources, estimatedTimeline, milestones, riskFactors, expectedOutputs, visibility, cloneCount, useCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        industry:
          type: string
        subIndustry:
          type: ["string", "null"]
        productCategory:
          type: ["string", "null"]
        specificProduct:
          type: ["string", "null"]
        description:
          type: string
        problemSolved:
          type: ["string", "null"]
        durationDays:
          type: integer
        requiredSkills:
          type: array
          items:
            type: string
        requiredTools:
          type: array
          items:
            type: string
        requiredResources:
          type: array
          items:
            type: string
        estimatedTimeline:
          type: string
        estimatedStartupCost:
          type: ["number", "null"]
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        dailyStructure:
          type: array
          items:
            type: object
        riskFactors:
          type: array
          items:
            type: string
        expectedOutputs:
          type: array
          items:
            type: string
        coverImageUrl:
          type: ["string", "null"]
        attachments:
          type: array
          items:
            type: string
        visibility:
          type: string
          enum: [draft, public, private]
        creatorId:
          type: ["integer", "null"]
        cloneCount:
          type: integer
        useCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        templateType:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: ["string", "null"]
          enum: [beginner, intermediate, advanced, null]
        followCount:
          type: integer
        saveCount:
          type: integer
        adoptionCount:
          type: integer
        forkCount:
          type: integer
        forkedFromId:
          type: ["integer", "null"]
        audioDescription:
          type: ["string", "null"]
    ExecutionInstance:
      type: object
      required: [id, ownerId, title, status, progressPercent, milestones, createdAt]
      properties:
        id:
          type: integer
        templateId:
          type: ["integer", "null"]
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        instanceType:
          type: string
        status:
          type: string
          enum: [active, paused, completed, abandoned]
        country:
          type: ["string", "null"]
        stateCity:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        targetEndDate:
          type: ["string", "null"]
        progressPercent:
          type: integer
        milestones:
          type: array
          items:
            type: object
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VentureTemplateInput:
      type: object
      required: [title, industry, description, durationDays, estimatedTimeline]
      properties:
        title:
          type: string
        industry:
          type: string
        subIndustry:
          type: string
        productCategory:
          type: string
        specificProduct:
          type: string
        description:
          type: string
        problemSolved:
          type: string
        durationDays:
          type: integer
        requiredSkills:
          type: array
          items:
            type: string
        requiredTools:
          type: array
          items:
            type: string
        requiredResources:
          type: array
          items:
            type: string
        estimatedTimeline:
          type: string
        estimatedStartupCost:
          type: number
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        dailyStructure:
          type: array
          items:
            type: object
        riskFactors:
          type: array
          items:
            type: string
        expectedOutputs:
          type: array
          items:
            type: string
        coverImageUrl:
          type: string
        attachments:
          type: array
          items:
            type: string
        visibility:
          type: string
          enum: [draft, public, private]
    Milestone:
      type: object
      required: [day, title, description]
      properties:
        day:
          type: integer
        title:
          type: string
        description:
          type: string
...